Auto Arc / Pure Velocity Base Test
The methods themselves work and the autonomous parser knows how to call the methods based on the script, but the scripts have not been changed and the methods have not been fully tested one right after the other due to lack of space. Maybe Andy or someone could mess around with it in the STEM center?